+++ lumin/nn/models/model.py
@@ -47,14 +47,14 @@ class Model(AbsModel):
             raise KeyError(key)
         raise ValueError(f'Expected string or int, recieved {key} of type {type(key)}')
         
-    def fit(self, batch_yielder:BatchYielder, callbacks:List[AbsCallback]) -> float:
+    def fit(self, batch_yielder:BatchYielder, callbacks:List[AbsCallback]=[]) -> float:
         self.model.train()
         self.stop_train = False
         losses = []
         for c in callbacks: c.on_epoch_begin(batch_yielder=batch_yielder)
 
         for x, y, w in batch_yielder:
-            for c in callbacks: c.on_batch_begin(inputs=x, targets=y, weights=w)
+            for c in callbacks: c.on_batch_begin()
             y_pred = self.model(x)
             loss = self.loss(weight=w)(y_pred, y) if w is not None else self.loss()(y_pred, y)
             losses.append(loss.data.item())
@@ -68,12 +68,14 @@ class Model(AbsModel):
         for c in callbacks: c.on_epoch_end(losses=losses)
         return np.mean(losses)
               
-    def evaluate(self, inputs:Tensor, targets:Tensor, weights:Optional[Tensor]=None) -> float:
+    def evaluate(self, inputs:Tensor, targets:Tensor, weights:Optional[Tensor]=None, callbacks:List[AbsCallback]=[]) -> float:
         self.model.eval()
         if 'multiclass' in self.objective: targets = targets.long().squeeze()
         else:                              targets = targets.float()
         y_pred = self.model(to_device(inputs.float()))
+        for c in callbacks: c.on_eval_begin(inputs=inputs, targets=targets, weights=weights)
         loss = self.loss(weight=to_device(weights))(y_pred, to_device(targets)) if weights is not None else self.loss()(y_pred, to_device(targets))
+        for c in callbacks: c.on_eval_end(loss=loss)        
         return loss.data.item()
 
     def predict_array(self, inputs:Union[np.ndarray, pd.DataFrame, Tensor, FoldYielder], as_np:bool=True) -> Union[np.ndarray, Tensor]: